Embedding ViewPortManager into ViewPort implementations
